+++ src/main/kotlin/com/abeade/plugin/figma/ImportDialogWrapper.kt
@@ -1,20 +1,31 @@
 package com.abeade.plugin.figma
 
 import com.abeade.plugin.figma.ui.ImportDialog
+import com.abeade.plugin.figma.ui.PreviewPanel
 import com.abeade.plugin.figma.ui.autocomplete.Autocomplete
 import com.abeade.plugin.figma.utils.EMPTY
 import com.abeade.plugin.figma.utils.containsAny
 import com.abeade.plugin.figma.utils.findFirstOf
+import com.abeade.plugin.figma.utils.isValidEntry
 import com.intellij.icons.AllIcons
 import com.intellij.ide.util.PropertiesComponent
+import com.intellij.notification.Notification
+import com.intellij.notification.NotificationType
+import com.intellij.notification.Notifications
 import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.openapi.ui.ValidationInfo
+import com.intellij.openapi.ui.popup.JBPopup
+import com.intellij.openapi.ui.popup.JBPopupFactory
 import com.intellij.ui.IdeBorderFactory
+import com.intellij.vcs.commit.NonModalCommitPanel.Companion.showAbove
 import java.awt.Desktop
+import java.awt.Dimension
 import java.awt.event.*
 import java.io.File
+import java.io.IOException
 import java.net.URI
 import java.util.zip.ZipFile
+import javax.imageio.ImageIO
 import javax.swing.*
 import javax.swing.event.DocumentEvent
 import javax.swing.event.DocumentListener
@@ -88,6 +99,27 @@ class ImportDialogWrapper(
             moreInfoLabel.addMouseListener(object : MouseAdapter() {
                 override fun mouseClicked(e: MouseEvent?) {
                     Desktop.getDesktop().browse(URI("https://github.com/abeade/figma-import-plugin"))
+                }
+            })
+            ldpiIconViewLabel.icon = AllIcons.General.InspectionsEye
+            mdpiIconViewLabel.icon = AllIcons.General.InspectionsEye
+            hdpiIconViewLabel.icon = AllIcons.General.InspectionsEye
+            xhdpiIconViewLabel.icon = AllIcons.General.InspectionsEye
+            xxhdpiIconViewLabel.icon = AllIcons.General.InspectionsEye
+            xxxhdpiIconViewLabel.icon = AllIcons.General.InspectionsEye
+            val previewTooltipText = "Click here to view this image"
+            ldpiIconViewLabel.toolTipText = previewTooltipText
+            mdpiIconViewLabel.toolTipText = previewTooltipText
+            hdpiIconViewLabel.toolTipText = previewTooltipText
+            xhdpiIconViewLabel.toolTipText = previewTooltipText
+            xxhdpiIconViewLabel.toolTipText = previewTooltipText
+            xxxhdpiIconViewLabel.toolTipText = previewTooltipText
+            previewLabel.icon = AllIcons.General.InspectionsEye
+            previewLabel.toolTipText = "Click here to view the images in this file"
+            previewLabel.isVisible = false
+            previewLabel.addMouseListener(object : MouseAdapter() {
+                override fun mouseClicked(e: MouseEvent?) {
+                    file?.let { FilePreviewDialogWrapper(it).show() }
                 }
             })
         }
@@ -242,17 +274,25 @@ class ImportDialogWrapper(
         updateLabelField(dialog.xxxhdpiLabel, dialog.xxxhdpiField)
         zipFilesList?.let {
             val skip = propertiesComponent.isTrueValue(SKIP_KEY)
-            updateIconField(dialog.ldpiIconLabel, dialog.ldpiField, it, skip, Density.LDPI)
-            updateIconField(dialog.mdpiIconLabel, dialog.mdpiField, it, skip, Density.MDPI)
-            updateIconField(dialog.hdpiIconLabel, dialog.hdpiField, it, skip, Density.HDPI)
-            updateIconField(dialog.xhdpiIconLabel, dialog.xhdpiField, it, skip, Density.XHDPI)
-            updateIconField(dialog.xxhdpiIconLabel, dialog.xxhdpiField, it, skip, Density.XXHDPI)
-            updateIconField(dialog.xxxhdpiIconLabel, dialog.xxxhdpiField, it, skip, Density.XXXHDPI)
+            updateIconField(dialog.ldpiIconLabel, dialog.ldpiIconViewLabel, dialog.ldpiField, it, skip, Density.LDPI)
+            updateIconField(dialog.mdpiIconLabel, dialog.mdpiIconViewLabel, dialog.mdpiField, it, skip, Density.MDPI)
+            updateIconField(dialog.hdpiIconLabel, dialog.hdpiIconViewLabel, dialog.hdpiField, it, skip, Density.HDPI)
+            updateIconField(dialog.xhdpiIconLabel, dialog.xhdpiIconViewLabel, dialog.xhdpiField, it, skip, Density.XHDPI)
+            updateIconField(dialog.xxhdpiIconLabel, dialog.xxhdpiIconViewLabel, dialog.xxhdpiField, it, skip, Density.XXHDPI)
+            updateIconField(dialog.xxxhdpiIconLabel, dialog.xxxhdpiIconViewLabel, dialog.xxxhdpiField, it, skip, Density.XXXHDPI)
+        } ?: run {
+            dialog.ldpiIconViewLabel.isVisible = false
+            dialog.mdpiIconViewLabel.isVisible = false
+            dialog.hdpiIconViewLabel.isVisible = false
+            dialog.xhdpiIconViewLabel.isVisible = false
+            dialog.xxhdpiIconViewLabel.isVisible = false
+            dialog.xxxhdpiIconViewLabel.isVisible = false
         }
     }
 
     private fun updateIconField(
         iconLabel: JLabel,
+        viewIconLabel: JLabel,
         field: JTextField,
         filesList: MutableList<String>,
         skipWhenNotExists: Boolean,
@@ -262,19 +302,65 @@ class ImportDialogWrapper(
         val suffix = field.text
         if (suffix.isBlank()) {
             iconLabel.icon = AllIcons.General.Warning
+            viewIconLabel.isVisible = false
             iconLabel.toolTipText = "Empty suffix. Density will be skipped"
         } else if (!destinationExists && skipWhenNotExists) {
             iconLabel.icon = AllIcons.General.Warning
+            viewIconLabel.isVisible = false
             iconLabel.toolTipText = "Density folder not found, resource will be skipped<br/>You can change this in plugin settings"
         } else {
-            if (filesList.any { it.substringBeforeLast('.').endsWith(suffix) }) {
+            val file = filesList.firstOrNull { it.substringBeforeLast('.').endsWith(suffix) }
+            if (file != null) {
                 iconLabel.icon = AllIcons.General.InspectionsOK
                 iconLabel.toolTipText = "Resource found"
+                viewIconLabel.isVisible = true
+                viewIconLabel.mouseListeners.forEach { viewIconLabel.removeMouseListener(it) }
+                viewIconLabel.addMouseListener(object : MouseAdapter() {
+                    override fun mouseClicked(e: MouseEvent?) {
+                        showPreviewPopup(file, viewIconLabel)
+                    }
+                })
             } else {
                 iconLabel.icon = AllIcons.General.Error
+                viewIconLabel.isVisible = false
                 iconLabel.toolTipText = "Resource not found"
             }
         }
+    }
+
+    private fun showPreviewPopup(fileName: String, target: JComponent): JBPopup? =
+        getBufferedImage(fileName)?.let { image ->
+            val previewPanel = PreviewPanel()
+            previewPanel.quickDrawPanel.setImage(image)
+            previewPanel.labelSize.text = "${image.width} x ${image.height}"
+            val popupBuilder = JBPopupFactory.getInstance().createComponentPopupBuilder(previewPanel.mainPanel, null)
+            popupBuilder.createPopup().apply {
+                size = Dimension(
+                    /* width = */ minOf(image.width, MAX_PREVIEW_SIZE) + WIDTH_PREVIEW_MARGIN,
+                    /* height = */ minOf(image.height, MAX_PREVIEW_SIZE) + HEIGHT_PREVIEW_MARGIN
+                )
+                showAbove(target)
+            }
+        }
+
+    private fun getBufferedImage(densityFile: String) = try {
+        file?.let { selectedZip ->
+            ZipFile(selectedZip).use { zipFile ->
+                zipFile.entries().toList().find { it.name.endsWith(densityFile) }?.let { entry ->
+                    ImageIO.read(zipFile.getInputStream(entry))
+                }
+            }
+        }
+    } catch (e: IOException) {
+        Notifications.Bus.notify(
+            Notification(
+                "Figma import",
+                "Error loading preview image",
+                "An error occurred processing loading image file $densityFile",
+                NotificationType.ERROR
+            )
+        )
+        null
     }
 
     private fun Density.getFolder(): String {
@@ -301,12 +387,10 @@ class ImportDialogWrapper(
         if (result == JFileChooser.APPROVE_OPTION) {
             file = fileDialog.selectedFile
             dialog.fileField.text = fileDialog.selectedFile.toString()
+            dialog.previewLabel.isVisible = true
             ZipFile(fileDialog.selectedFile).use { zipFile ->
                 zipFilesList = zipFile.entries().asSequence()
-                    .filter {
-                        !it.isDirectory && (it.name.endsWith(".png", true) ||
-                                it.name.endsWith(".jpg", true))
-                    }
+                    .filter { it.isValidEntry() }
                     .fold(mutableListOf()) { list, entry -> list.apply { add(File(entry.name).name) } }
             }
             zipFilesList?.let {
@@ -354,5 +438,9 @@ class ImportDialogWrapper(
 
         private const val QUALIFIER_SEPARATOR = "-"
         private const val FOLDER_DRAWABLE = "drawable$QUALIFIER_SEPARATOR"
+
+        private const val MAX_PREVIEW_SIZE = 400
+        private const val WIDTH_PREVIEW_MARGIN = 4
+        private const val HEIGHT_PREVIEW_MARGIN = 26
     }
 }