+++ src/main/kotlin/com/abeade/plugin/figma/utils/ZipEntryExtensions.kt
@@ -1,0 +1,6 @@
+package com.abeade.plugin.figma.utils
+
+import java.util.zip.ZipEntry
+
+fun ZipEntry.isValidEntry() = !isDirectory && (name.endsWith(".png", true) ||
+    name.endsWith(".jpg", true))